+++ lib/cucumber/formatter/errors.rb
@@ -1,0 +1,6 @@
+module Cucumber
+  module Formatter
+    class TestCaseUnknownError < StandardError; end
+    class TestStepUnknownError < StandardError; end
+  end
+end